chore: move durations into a single dir, and claify range duration vs windowPeriod duration. Have notebooks use the updated windowVar methods.
