+++ src/shared/utils/duration/getDurationFromAST.test.ts
